fix(strategies): move to openllm
Strategies shouldn't be a part of openllm-core
